suneung-nonfiction/requirements.txt
pymupdf==1.24.9
kss==6.0.4
scikit-learn==1.5.2
numpy>=1.26
pandas>=2.2
pydantic==2.8.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
openai>=1.54.0